simplify files; implement seqFISH
