+++ scvi/dataset/__init__.py
@@ -2,15 +2,14 @@ from .brain_large import BrainLargeDataset
 from .cortex import CortexDataset
 from .dataset import GeneExpressionDataset
 from .synthetic import SyntheticDataset
-from .retina import RetinaDataset
 from .cite_seq import CiteSeqDataset, CbmcDataset
-from .pbmc import PbmcDataset
-from .brain_small import BrainSmallDataset
 from .hemato import HematoDataset
-from .loom import LoomDataset
+from .loom import LoomDataset, RetinaDataset
 from .dataset10X import Dataset10X
 from .anndata import AnnDataset
 from .csv import CsvDataset
+from .dataset10X import PbmcDataset, BrainSmallDataset
+from .seqfish import SeqfishDataset
 
 __all__ = ['SyntheticDataset',
            'CortexDataset',
@@ -25,4 +24,5 @@ __all__ = ['SyntheticDataset',
            'LoomDataset',
            'AnnDataset',
            'CsvDataset',
-           'Dataset10X']
+           'Dataset10X',
+           'SeqfishDataset']
