+++ scvi/dataset/loom.py
@@ -62,3 +62,10 @@ class LoomDataset(GeneExpressionDataset):
 
         print("Finished preprocessing dataset")
         return data, gene_names, labels, cell_batches
+
+
+class RetinaDataset(LoomDataset):
+    def __init__(self, save_path='data/'):
+        super(RetinaDataset, self).__init__(filename='retina.loom',
+                                            save_path=save_path,
+                                            url='https://github.com/YosefLab/scVI-data/raw/master/retina.loom')
